:sparkles: Add a global guild count
The guild count we can get from the db only includes guilds that are
or have actively used Titi, whereas we can get this information directly
from Discord itself to get a full count.

Updated the stats command to also return this information as well
as outputting it to logs.

Added some event handlers just to keep track of newly added / removed
guilds so the count doesn't get too stale.